+++ docassemble_base/docassemble/base/astparser.py
@@ -1,6 +1,6 @@
 import ast
 import re
-import sys
+#import sys
 
 fix_assign = re.compile(r'\.(\[[^\]]*\])')
 valid_variable_match = re.compile(r'^[^\d][A-Za-z0-9\_]*$')
@@ -55,8 +55,10 @@ class myvisitnode(ast.NodeVisitor):
         self.depth -= 1
     def visit_Call(self, node):
         self.calls.add(node.func)
-        if hasattr(node.func, 'id') and node.func.id in ['showif', 'showifdef', 'value', 'defined'] and len(node.args) and node.args[0].__class__.__name__ == 'Str' and hasattr(node.args[0], 's') and re.search(r'^[^\d]', node.args[0].s) and not re.search(r'[^A-Za-z0-9\.\"\'\[\] ]', node.args[0].s):
+        if hasattr(node.func, 'id') and node.func.id in ['showif', 'showifdef', 'value', 'defined'] and len(node.args) and node.args[0].__class__.__name__ == 'Str' and hasattr(node.args[0], 's') and re.search(r'^[^\d]', node.args[0].s) and not re.search(r'[^A-Z_a-z0-9\.\"\'\[\] ]', node.args[0].s):
             self.names[node.args[0].s] = 1
+        if hasattr(node.func, 'id') and node.func.id in ['define'] and len(node.args) and node.args[0].__class__.__name__ == 'Str' and hasattr(node.args[0], 's') and re.search(r'^[^\d]', node.args[0].s) and not re.search(r'[^A-Z_a-z0-9\.\"\'\[\] ]', node.args[0].s):
+            self.targets[node.args[0].s] = 1
         ast.NodeVisitor.generic_visit(self, node)
     def visit_Subscript(self, node):
         if node not in self.calls: